pyproject.toml
[tool.poetry]
name = "ole-tools"
version = "0.1.0"
description = ""
authors = ["Unknown"]
packages = [{ include = "ole" }]

[tool.poetry.scripts]
ole_measure = 'ole.scripts.measure_display:main'
ole_analyze = 'ole.scripts.analyze_display_measurements:main'
ole_anonymize = 'ole.scripts.strip_metadata:main'

[tool.poetry.dependencies]
requests = "^2.31.0"
python = "~3.12"
colour-science = "^0.4.4"
colour-datasets = "^0.2.5"
colour-specio = "^0.3.0"
numpy = "^1.25.0"
scipy = "^1.10.1"
xxhash = "^3.2.0"
matplotlib = "^3.7.1"
scikit-learn = "^1.2.2"
pandas = "^2.2.0"
opencv-python = "^4.8.1.78"
pymoo = "^0.6.1.1"
telnetlib3 = "^2.0.4"
exscript = "^2.6.28"
echion = "^0.3.0"

[tool.poetry.dev-dependencies]
jupyter = "*"
black = "*"
pytest = "*"
pytest-xdist = "*"
pyright = "*"
ruff = "*"
flynt = "*"
isort = "*"
ipykernel = "*"
networkx = ">= 2.7, < 3"
opencolorio = ">= 2, < 3"
tqdm = "*"

colour-science = { path = "colour-science", develop = true }
colour-datasets = { path = "colour-datasets", develop = true }
colour-specio = { path = "colour-specio", develop = true }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.codespell]
ignore-words-list = 'co-ordinates,exitance,hart,ist'
skip = 'BIBLIOGRAPHY.bib,CONTRIBUTORS.rst'

[tool.flynt]
line_length = 999

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
skip_glob = ["colour/**/__init__.py"]
split_on_trailing_comma = true
use_parentheses = true


[tool.pyright]
reportMissingImports = false
reportMissingModuleSource = false
reportUnboundVariable = false
reportUnnecessaryCast = true
reportUnnecessaryTypeIgnorComment = true
reportUnsupportedDunderAll = false
reportUnusedExpression = false

[tool.pytest.ini_options]
addopts = "-n auto --dist=loadscope --durations=5 --import-mode=importlib"
filterwarnings = [
    "ignore::RuntimeWarning",
    "ignore::pytest.PytestCollectionWarning",
    "ignore::colour.utilities.ColourWarning",
    "ignore::colour.utilities.ColourRuntimeWarning",
    "ignore::colour.utilities.ColourUsageWarning",
    "ignore:Implicit None on return values is deprecated:DeprecationWarning",
    "ignore:Jupyter is migrating its paths:DeprecationWarning",
    "ignore:the imp module is deprecated:DeprecationWarning",
    "ignore:Method Nelder-Mead does not use gradient information:RuntimeWarning",
    "ignore:More than 20 figures have been opened:RuntimeWarning",
    "ignore:divide by zero encountered:RuntimeWarning",
    "ignore:invalid value encountered in:RuntimeWarning",
    "ignore:overflow encountered in:RuntimeWarning",
    "ignore:Matplotlib is currently using agg:UserWarning",
    "ignore:override the edgecolor or facecolor properties:UserWarning",
]

[tool.ruff]
target-version = "py39"
line-length = 88
select = [
    "A",   # flake8-builtins
    "ARG", # flake8-unused-arguments
    # "ANN", # flake8-annotations
    "B", # flake8-bugbear
    # "BLE", # flake8-blind-except
    "C4", # flake8-comprehensions
    # "C90", # mccabe
    # "COM", # flake8-commas
    "DTZ", # flake8-datetimez
    "D",   # pydocstyle
    "E",   # pydocstyle
    # "ERA", # eradicate
    # "EM", # flake8-errmsg
    "EXE", # flake8-executable
    "F",   # flake8
    # "FBT", # flake8-boolean-trap
    "G",   # flake8-logging-format
    "I",   # isort
    "ICN", # flake8-import-conventions
    "INP", # flake8-no-pep420
    "ISC", # flake8-implicit-str-concat
    "N",   # pep8-naming
    # "PD", # pandas-vet
    "PIE", # flake8-pie
    "PGH", # pygrep-hooks
    "PL",  # pylint
    # "PT", # flake8-pytest-style
    # "PTH", # flake8-use-pathlib [Enable]    "Q",   # flake8-quotes
    "RET", # flake8-return
    "RUF", # Ruff
    "S",   # flake8-bandit
    "SIM", # flake8-simplify
    "T10", # flake8-debugger
    "T20", # flake8-print
    # "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "TRY", # tryceratops
    "UP",  # pyupgrade
    "W",   # pydocstyle
    "YTT", # flake8-2020
]
ignore = [
    "B008",
    "B905",
    "D104",
    "D200",
    "D202",
    "D205",
    "D301",
    "D400",
    "I001",
    "N801",
    "N802",
    "N803",
    "N806",
    "N813",
    "N815",
    "N816",
    "PGH003",
    "PIE804",
    "PLE0605",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "PLR2004",
    "RET504",
    "RET505",
    "RET506",
    "RET507",
    "RET508",
    "TRY003",
    "TRY300",
]
typing-modules = ["colour.hints"]
fixable = ["B", "C", "E", "F", "PIE", "RUF", "SIM", "UP", "W"]

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.ruff.per-file-ignores]
"colour/examples/*" = ["INP", "T201", "T203"]
"docs/*" = ["INP"]
"tasks.py" = ["INP"]
"utilities/*" = ["EXE001", "INP"]
"utilities/unicode_to_ascii.py" = ["RUF001"]

[tool.ruff.format]
docstring-code-format = true
